code/chapter06/camel_demo/pyproject.toml
[project]
name = "camel-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Alwaysungry"}
]
requires-python = ">=3.13, <3.15"
dependencies = [
    "colorama (>=0.4.6,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
